remove open room button once a room has been created to avoid spam
