+++ backend-fe/socketServerStores/roomsLive.js
@@ -28,20 +28,7 @@ const getRoomParticipants = (room) => {
   return roomsLive[room].participants;
 };
 
-const sendRoomMessage = (message) => {
-  console.log("message:");
-  console.log(message);
-  return message;
-};
-
-const getRoomMessages = () => {
-  console.log("sending");
-  // return roomMessages;
-};
-
 module.exports = {
   joinRoom,
   getRoomParticipants,
-  sendRoomMessage,
-  getRoomMessages,
 };
