+++ raynian-fe/src/components/rooms/RoomButton.jsx
@@ -1,18 +1,23 @@
 import React, { useContext } from "react";
 import { ModalContext } from "../../context/ModalContext";
+import { useSelector } from "react-redux";
 
 const RoomButton = () => {
+  const { roomId } = useSelector((state) => state.room);
   const { toggleJoinRoomPrompt, toggleOpenRoomPrompt } =
     useContext(ModalContext);
 
   return (
     <div className="mt-4">
-      <button
-        onClick={toggleOpenRoomPrompt}
-        className="ml-2 rounded-md bg-blue-500 px-4 py-2 text-white"
-      >
-        Open your room
-      </button>
+      {roomId == null && (
+        <button
+          onClick={toggleOpenRoomPrompt}
+          className="ml-2 rounded-md bg-blue-500 px-4 py-2 text-white"
+        >
+          Open your room
+        </button>
+      )}
+
       <button
         onClick={toggleJoinRoomPrompt}
         className="ml-2 rounded-md bg-blue-500 px-4 py-2 text-white"
